update blog cards to use styled components
also rename the file to use camel case
